--- admin/src/main/java/com/kernel/admin/AdminApplication.java
@@ -1,13 +1,0 @@
-package com.kernel.admin;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@SpringBootApplication
-public class AdminApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AdminApplication.class, args);
-    }
-
-}
