--- admin/src/test/java/com/kernel/admin/AdminApplicationTests.java
@@ -1,14 +1,0 @@
-package com.kernel.admin;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class AdminApplicationTests {
-
-    @Test
-    void contextLoads() {
-        // TODO: 테스트 작성 예정
-    }
-
-}
