+++ api/src/main/java/com/kernel/app/ApiApplication.java
@@ -1,0 +1,17 @@
+package com.kernel.app;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
+import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
+
+@SpringBootApplication(scanBasePackages = "com.kernel")
+@EnableJpaRepositories(basePackages = "com.kernel")
+@EntityScan(basePackages = "com.kernel")
+public class ApiApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ApiApplication.class, args);
+    }
+
+}
