+++ api/src/test/java/com/kernel/app/ApiApplicationTests.java
@@ -1,0 +1,13 @@
+package com.kernel.app;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ApiApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
