--- app/src/main/java/com/kernel/app/AppApplication.java
@@ -1,17 +1,0 @@
-package com.kernel.app;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.domain.EntityScan;
-import org.springframework.data.jpa.repository.config.EnableJpaRepositories;
-
-@SpringBootApplication(scanBasePackages = "com.kernel")
-@EnableJpaRepositories(basePackages = "com.kernel")
-@EntityScan(basePackages = "com.kernel")
-public class AppApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(AppApplication.class, args);
-    }
-
-}
