+++ build.gradle
@@ -19,4 +19,8 @@ subprojects {
 		sourceCompatibility = JavaVersion.VERSION_17
 		targetCompatibility = JavaVersion.VERSION_17
 	}
+
+	tasks.test {
+		useJUnitPlatform()
+	}
 }
