+++ common/build.gradle
@@ -24,9 +24,9 @@ dependencies {
     //implementation 'org.springframework.boot:spring-boot-starter-security'
 
     // JWT
-    //implementation 'io.jsonwebtoken:jjwt-api:0.11.5'
-    //runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.11.5'
-    //runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.11.5'
+    //implementation 'io.jsonwebtoken:jjwt-api:0.12.3'
+    //runtimeOnly 'io.jsonwebtoken:jjwt-impl:0.12.3'
+    //runtimeOnly 'io.jsonwebtoken:jjwt-jackson:0.12.3' // JSON 처리용
 
     // Lombok (개발 편의)
     compileOnly 'org.projectlombok:lombok'
@@ -39,6 +39,8 @@ dependencies {
     implementation 'org.springframework.boot:spring-boot-starter-validation'
     testImplementation 'org.junit.jupiter:junit-jupiter:5.10.0'
 
+    // mysql
+    implementation 'com.mysql:mysql-connector-j'
 
 }
