--- common/src/main/java/com/kernel/common/CommonApplication.java
@@ -1,13 +1,0 @@
-package com.kernel.common;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@SpringBootApplication
-public class CommonApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CommonApplication.class, args);
-    }
-
-}
