--- customer/src/main/java/com/kernel/customer/CustomerApplication.java
@@ -1,13 +1,0 @@
-package com.kernel.customer;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@SpringBootApplication
-public class CustomerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CustomerApplication.class, args);
-    }
-
-}
