--- customer/src/test/java/com/kernel/customer/CustomerApplicationTests.java
@@ -1,14 +1,0 @@
-package com.kernel.customer;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class CustomerApplicationTests {
-
-    @Test
-    void contextLoads() {
-        // TODO: 테스트 작성 예정
-    }
-
-}
