--- manager/src/main/java/com/kernel/manager/ManagerApplication.java
@@ -1,13 +1,0 @@
-package com.kernel.manager;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@SpringBootApplication
-public class ManagerApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ManagerApplication.class, args);
-    }
-
-}
