--- manager/src/test/java/com/kernel/manager/ManagerApplicationTests.java
@@ -1,14 +1,0 @@
-package com.kernel.manager;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ManagerApplicationTests {
-
-    @Test
-    void contextLoads() {
-        // TODO: 테스트 작성 예정
-    }
-
-}
