--- matching/src/main/java/com/kernel/matching/MatchingApplication.java
@@ -1,13 +1,0 @@
-package com.kernel.matching;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@SpringBootApplication
-public class MatchingApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(MatchingApplication.class, args);
-    }
-
-}
