--- matching/src/test/java/com/kernel/matching/MatchingApplicationTests.java
@@ -1,14 +1,0 @@
-package com.kernel.matching;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class MatchingApplicationTests {
-
-    @Test
-    void contextLoads() {
-        // TODO: 테스트 작성 예정
-    }
-
-}
