--- payment/src/main/java/com/kernel/payment/PaymentApplication.java
@@ -1,13 +1,0 @@
-package com.kernel.payment;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-//@SpringBootApplication
-public class PaymentApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(PaymentApplication.class, args);
-    }
-
-}
