--- payment/src/test/java/com/kernel/payment/PaymentApplicationTests.java
@@ -1,14 +1,0 @@
-package com.kernel.payment;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class PaymentApplicationTests {
-
-    @Test
-    void contextLoads() {
-        // TODO: 테스트 작성 예정
-    }
-
-}
