--- reservation/src/test/java/com/kernel/reservation/ReservationApplicationTests.java
@@ -1,14 +1,0 @@
-package com.kernel.reservation;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class ReservationApplicationTests {
-
-    @Test
-    void contextLoads() {
-        // TODO: 테스트 작성 예정
-    }
-
-}
